Refactor code to push img retrieve client side
