+++ server/hpd_sensors.py
@@ -275,123 +275,3 @@ class MyAudio(threading.Thread):
             writer = threading.Thread(target=self.write_to_file, args = (f_path, self.frames))
             writer.start()
             writer.join()
-                
-
-class MyPhoto(threading.Thread):
-    def __init__(self, img_root, stream_type, debug):
-        threading.Thread.__init__(self)
-        # self.setDaemon(True)
-        print("Initializing MyPhoto class")
-        self.img_root = img_root
-        self.debug = debug
-        self.img_root_date = os.path.join(self.img_root, datetime.now().strftime("%Y-%m-%d"))
-        self.pi_ip_address = 'localhost'
-        self.stream_type = stream_type
-        self.video_status = False
-        self.create_root_img_dir()
-        self.connect_to_video()
-        self.start()
-
-    def create_root_img_dir(self):
-        if not os.path.isdir(self.img_root):
-            os.makedirs(self.img_root)
-
-    def connect_to_video(self):
-        # Select the stream type based on that specified in server.conf
-        if self.stream_type == "mjpeg":
-            stream_path = "http://" + self.pi_ip_address + ":8080/stream/video.mjpeg"
-        elif self.stream_type == "h264":
-            stream_path = "http://" + self.pi_ip_address + ":8080/stream/video.h264"
-        elif self.stream_type == "jpeg":
-            stream_path = "http://" + self.pi_ip_address + ":8080/stream/video.jpeg"
-        
-        # Attempt to start the video stream
-        self.cam = WebcamVideoStream(stream_path).start()
-
-        self.img_restart_attempts = 0
-        # Keep attempting to open the video stream until it is opened
-        while not self.cam.stream.isOpened():
-            self.cam = WebcamVideoStream(stream_path).start()
-            self.video_status = False
-            self.img_restart_attempts += 1
-            logging.warning("Unable to connect to video")
-            if self.debug:
-                print('Unable to connect to video')
-            if self.img_restart_attempts >= 5:
-                subprocess.run("sudo reboot", shell = True)
-                # subprocess.run("sudo service uv4l_raspicam restart", shell = True)
-                # time.sleep(5)
-                # self.img_restart_attempts = 0
-                # subprocess.run("sudo service hpd_mobile restart", shell = True)
-
-            time.sleep(1)
-
-        # Set the video status to true
-        self.video_status = True
-        logging.info("Connected to video stream")
-        if self.debug:
-            print('Connected to video stream')
-
-    def img_dir_update(self):
-        # This function is run in a separate thread to continuously create a new directory for each day, and for each minute.
-        while 1:
-            date_dir = os.path.join(self.img_root, datetime.now().strftime("%Y-%m-%d"))
-            if not os.path.isdir(date_dir):
-                os.makedirs(date_dir)
-            self.img_root_date = date_dir
-            
-            min_dir = os.path.join(self.img_root_date, datetime.now().strftime("%H%M"))
-            if not os.path.isdir(min_dir):
-                os.makedirs(min_dir)
-            self.img_dir = min_dir
-
-    def run(self):
-        dir_create = threading.Thread(target=self.img_dir_update)
-        dir_create.start()
-        
-        # Wait for self.img_dir to exist
-        time.sleep(1)
-        while 1:
-            f_name = datetime.now().strftime("%Y-%m-%d %H%M%S_photo.png")
-            f_path = os.path.join(self.img_dir, f_name)
-
-            # Only capture a photo if it doesn't already exist
-            if not os.path.isfile(f_path):
-                if not len(os.listdir(self.img_dir)) >= 60:
-                    img = False
-                    img = self.cam.read()
-                    if type(img) is not np.ndarray:
-                        logging.warning('Camera read did not return image.  Attempting to restart video connection')
-                        if self.debug:
-                            print('Camera read did not return image.  Attempting to restart video connection')
-                        self.video_status = False
-                        try:
-                            self.cam.close()
-                        except Exception as e:
-                            logging.warning('Unable to close cam.  Potentially closed.  Exception: {}'.format(e))
-                        self.connect_to_video()
-
-                    elif type(img) is np.ndarray:
-                        try:
-                            # Convert image to greyscale
-                            img = cv2.cvtColor(img, cv2.COLOR_RGB2GRAY)
-
-                            # Write to disk
-                            # This is closing the image file???
-                            # look into /var/log/messages grep for oom killer
-                            cv2.imwrite(f_path, img)
-                            if datetime.now().second == 0:
-                                logging.info("Created file: {}".format(f_path))
-
-                        except Exception as e:
-                            logging.warning("Unable to convert to grayscale and write to disk.  Error: {}.  File: {}\tAttempting to restart video connection".format(e, f_name))
-                            if self.debug:
-                                print("Unable to convert to grayscale and write to disk.  Error: {}.  File: {}\tAttempting to restart video connection".format(e, f_name))
-                            # logging.info("Attempting to restart video connection")
-                            self.video_status = False
-                            try:
-                                self.cam.close()
-                            except Exception as e:
-                                logging.warning('Unable to close cam.  Potentially closed.  Exception: {}'.format(e))
-                            self.connect_to_video()
-                            # logging.CRITICAL("Unable to convert to grayscale and write to disk.  Error: {}.  File: {}".format(e, fname))
